before merging malka's

+++ src/main/java/com/ti/surveyserver/repository/SurveyShellRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.Repository;
 
 import java.math.BigInteger;
 
-public interface SurveyShellRepository extends Repository<SurveyShell, Long> {
+public interface SurveyShellRepository extends Repository<SurveyShell, BigInteger> {
     SurveyShell findOneById(BigInteger id);
 //    SurveyShell save(SurveyShell customer);
